+++ src/BenchmarkDotNet/ConsoleArguments/ListBenchmarks/IBenchmarkCasesPrinter.cs
@@ -4,6 +4,6 @@ namespace BenchmarkDotNet.ListBenchmarks
 {
     internal interface IBenchmarkCasesPrinter
     {
-        void Print(IEnumerable<string> testName);
+        void Print(IEnumerable<string> testNames);
     }
 }
